feat(ud): Phase 5 - one Vite process
