+++ local-testing-project/.yarnrc.yml
@@ -1,15 +1,12 @@
-# Yarn's manifest file. You can configure yarn here.
-# See https://yarnpkg.com/configuration/yarnrc.
+approvedGitRepositories:
+  - '**'
 
-# For `node_modules` (see `nodeLinker` below), this is almost always the preferred option.
 compressionLevel: 0
 
 enableGlobalCache: true
 
-# Lets yarn use hardlinks inside `node_modules` to dedupe packages.
-# For a more pnpm-like experience, consider `hardlinks-global` where hardlinks point to a global store.
+enableScripts: true
+
 nmMode: hardlinks-local
 
-# How to install Node packages.
-# Heads up: right now, Cedar expects this to be `node-modules`.
 nodeLinker: node-modules
