+++ packages/vite/src/cedar-unified-dev.ts
@@ -1,12 +1,31 @@
 #!/usr/bin/env node
+import type { IncomingMessage, ServerResponse } from 'node:http'
+
+import { createServerAdapter } from '@whatwg-node/server'
 import { createServer } from 'vite'
 import yargsParser from 'yargs-parser'
 
 import { getPaths, getConfig } from '@cedarjs/project-config'
 
-import { startApiDevServer } from './apiDevServer.js'
+import { startApiDevMiddleware } from './apiDevMiddleware.js'
+
+function isViteInternalRequest(url: string): boolean {
+  return (
+    url.startsWith('/@') ||
+    url.startsWith('/__vite') ||
+    url.startsWith('/__hmr') ||
+    url.includes('?import') ||
+    url.includes('?t=') ||
+    url.includes('?v=')
+  )
+}
 
 const startUnifiedDevServer = async () => {
+  // Signal to Cedar plugins (e.g. cedarWaitForApiServer) that we're running
+  // in unified-dev mode so they can skip behaviours that assume a separate
+  // API listener.
+  process.env.__CEDAR_UNIFIED_DEV = 'true'
+
   const rwPaths = getPaths()
   const cedarConfig = getConfig()
   const configFile = rwPaths.web.viteConfig
@@ -19,7 +38,7 @@ const startUnifiedDevServer = async () => {
     force: forceOptimize,
     debug,
     port: portArg,
-    apiPort: apiPortArg,
+    apiPort: _apiPortArg,
     _: _positional,
     ...serverArgs
   } = yargsParser(process.argv.slice(2), {
@@ -29,12 +48,12 @@ const startUnifiedDevServer = async () => {
 
   const webPort =
     (portArg as number | undefined) ?? cedarConfig.web.port ?? 8910
-  const apiPort =
-    (apiPortArg as number | undefined) ?? cedarConfig.api.port ?? 8911
 
-  // Start the API dev server (Vite SSR + Fastify) first so it's ready
-  // before the web dev server tries to proxy requests to it.
-  const { close: closeApi } = await startApiDevServer(apiPort)
+  // Start the API dev middleware (Vite SSR, no separate HTTP listener).
+  // API requests will be handled inline via the web Vite dev server's
+  // middleware pipeline.
+  const { close: closeApi, handler: apiHandler } = await startApiDevMiddleware()
+  const apiAdapter = createServerAdapter(apiHandler)
 
   const devServer = await createServer({
     configFile,
@@ -49,6 +68,75 @@ const startUnifiedDevServer = async () => {
       ...serverArgs,
     },
     logLevel: debug ? 'info' : undefined,
+    plugins: [
+      {
+        name: 'cedar-api-middleware',
+        apply: 'serve',
+        configureServer(server) {
+          const apiUrl = cedarConfig.web.apiUrl.replace(/\/$/, '')
+          const apiGqlUrl = cedarConfig.web.apiGraphQLUrl ?? apiUrl + '/graphql'
+
+          function isApiRequest(url: string): boolean {
+            return (
+              url === apiUrl ||
+              url.startsWith(apiUrl + '/') ||
+              url.startsWith(apiUrl + '?') ||
+              url === apiGqlUrl ||
+              url.startsWith(apiGqlUrl + '/') ||
+              url.startsWith(apiGqlUrl + '?')
+            )
+          }
+
+          server.middlewares.use(
+            async (
+              req: IncomingMessage,
+              res: ServerResponse,
+              next: () => void,
+            ) => {
+              const url = req.url ?? '/'
+
+              if (isViteInternalRequest(url)) {
+                return next()
+              }
+
+              if (!isApiRequest(url)) {
+                return next()
+              }
+
+              try {
+                await apiAdapter(req, res)
+              } catch (err) {
+                console.error(
+                  '[cedar-api-middleware] Error handling API request:',
+                  err,
+                )
+
+                if (!res.headersSent) {
+                  res.writeHead(500, { 'Content-Type': 'application/json' })
+                }
+
+                res.end(
+                  JSON.stringify(
+                    {
+                      errors: [
+                        {
+                          message:
+                            err instanceof Error
+                              ? err.message
+                              : 'Internal Server Error',
+                        },
+                      ],
+                    },
+                    null,
+                    2,
+                  ),
+                )
+              }
+            },
+          )
+        },
+      },
+    ],
   })
 
   await devServer.listen()
