+++ packages/vite/src/index.ts
@@ -33,7 +33,6 @@ export { cedarTransformJsAsJsx } from './plugins/vite-plugin-jsx-loader.js'
 export { cedarMergedConfig } from './plugins/vite-plugin-merged-config.js'
 export { cedarSwapApolloProvider } from './plugins/vite-plugin-swap-apollo-provider.js'
 export { cedarUniversalDeployPlugin } from './plugins/vite-plugin-cedar-universal-deploy.js'
-export { cedarDevDispatcherPlugin } from './plugins/vite-plugin-cedar-dev-dispatcher.js'
 export { cedarWaitForApiServer } from './plugins/vite-plugin-cedar-wait-for-api-server.js'
 
 type PluginOptions = {
